feat:retrieve data from the form and store it in the blockchain
